[wip]: different approach for filenames
This is a good way in languages with types that prevent you from
shooting yourself in the foot. Not so much when you can just cast a
number to the aliased type. But, adding a more complex enumerated type
prevents the nice `const`-ness of it.

Probably gonna throw this away.